@@ -6,7 +6,7 @@
 
     <groupId>org.monarchinitiative.fenominal</groupId>
     <artifactId>fenominal</artifactId>
-    <version>0.2.0-SNAPSHOT</version>
+    <version>0.3.5</version>
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
@@ -25,12 +25,10 @@
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-        <java.version>16</java.version>
+        <java.version>17</java.version>
         <maven.compiler.source>${java.version}</maven.compiler.source>
         <maven.compiler.target>${java.version}</maven.compiler.target>
         <javafx.version>${java.version}</javafx.version>
-
-        <hpotextmining.version>0.2.8</hpotextmining.version>
         <phenol.version>1.6.3</phenol.version>
         <phenopacket.version>1.0.0</phenopacket.version>
         <spring.version>4.3.8.RELEASE</spring.version>
@@ -90,11 +88,6 @@
                 <groupId>net.rgielen</groupId>
                 <artifactId>javafx-weaver-spring-boot-starter</artifactId>
                 <version>1.3.0</version>
-            </dependency>
-            <dependency>
-                <groupId>org.monarchinitiative.hpotextmining</groupId>
-                <artifactId>hpotextmining-gui</artifactId>
-                <version>${hpotextmining.version}</version>
             </dependency>
         </dependencies>
     </dependencyManagement>
