@@ -52,8 +52,7 @@ public class SimpleHpoTerm {
 
     @Override
     public boolean equals(Object obj) {
-        if (! (obj instanceof SimpleHpoTerm)) return false;
-        SimpleHpoTerm that = (SimpleHpoTerm) obj;
+        if (! (obj instanceof SimpleHpoTerm that)) return false;
         return this.id.equals(that.id) && this.name.equals(that.name);
     }
 
