@@ -51,16 +51,15 @@ public class FenominalApplication extends Application {
      */
     @Override
     public void stop() throws Exception {
-        super.stop();
         // save properties
         final Properties pgProperties = applicationContext.getBean("pgProperties", Properties.class);
         final Path configFilePath = applicationContext.getBean("configFilePath", Path.class);
         try (OutputStream os = Files.newOutputStream(configFilePath)) {
             pgProperties.store(os, "Fenominal properties");
         }
-        Platform.exit();
-        // close the context
+        super.stop();
         applicationContext.close();
+        Platform.exit();
     }
 
 
