@@ -1,11 +1,13 @@
 package org.monarchinitiative.fenominal.gui;
 
+import org.monarchinitiative.fenominal.core.MinedTerm;
+import org.monarchinitiative.fenominal.core.TermMiner;
 import org.monarchinitiative.fenominal.core.corenlp.MappedSentencePart;
 import org.monarchinitiative.fenominal.core.lexical.LexicalResources;
 import org.monarchinitiative.fenominal.core.textmapper.ClinicalTextMapper;
 import org.monarchinitiative.fenominal.core.SimpleMinedTerm;
-import org.monarchinitiative.hpotextmining.core.miners.MinedTerm;
-import org.monarchinitiative.hpotextmining.core.miners.TermMiner;
+
+
 import org.monarchinitiative.phenol.ontology.data.Ontology;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
