@@ -4,8 +4,10 @@ import org.monarchinitiative.fenominal.gui.OptionalResources;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.core.io.Resource;
 
 import java.io.File;
 import java.io.IOException;
@@ -21,7 +23,14 @@ public class FenominalConfig {
 
     public static final String CONFIG_FILE_BASENAME = "fenominal.properties";
 
-
+    @Value("classpath:/fxml/Main.fxml")
+    private Resource mainFxmResource;
+    @Value("classpath:/fxml/Configure.fxml")
+    private Resource configureFxmResource;
+    @Value("classpath:/fxml/OntologyTree.fxml")
+    private Resource ontologyTreeFxmResource;
+    @Value("classpath:/fxml/Present.fxml")
+    private Resource presentFxmResource;
     @Bean
     public OptionalResources optionalResources() {
         return new OptionalResources();
@@ -91,21 +100,11 @@ public class FenominalConfig {
     }
 
 
-//    @Bean("appNameVersion")
-//    String appNameVersion(String appVersion, String appName) {
-//        return String.format("%s : %s", appName, appVersion);
-//    }
-//
-//
-//    @Bean("appVersion")
-//    String appVersion() {
-//        // this property is set in FenominalApplication#init()
-//        return System.getProperty(FenominalApplication.FENOMINAL_VERSION_PROP_KEY);
-//    }
-//
-//    @Bean("appName")
-//    String appName() {
-//        // this property is set in FenominalApplication#init()
-//        return System.getProperty(FenominalApplication.FENOMINAL_NAME_KEY);
-//    }
+
+
+
+
+
+
+
 }
