@@ -52,7 +52,7 @@ public class Platform {
                 Alert a = new Alert(AlertType.ERROR);
                 a.setTitle("Find gui config dir");
                 a.setHeaderText(null);
-                a.setContentText(String.format("Unrecognized platform. %s", platform.toString()));
+                a.setContentText(String.format("Unrecognized platform. %s", platform));
                 a.showAndWait();
                 return null;
         }
@@ -133,7 +133,7 @@ public class Platform {
         OSX("Os X"),
         UNKNOWN("Unknown");
 
-        private String name;
+        private final String name;
 
         CurrentPlatform(String n) {this.name = n; }
 
