@@ -1,19 +1,15 @@
 package org.monarchinitiative.fenominal.gui.model;
 
-import org.monarchinitiative.hpotextmining.gui.controller.Main;
+import org.monarchinitiative.fenominal.gui.hpotextminingwidget.PhenotypeTerm;
 import org.monarchinitiative.phenol.ontology.data.Term;
 
-import java.util.Objects;
-
-public class FenominalTerm implements Comparable<FenominalTerm> {
-
-    private final Term term;
-    private final boolean observed;
-
-    private FenominalTerm(Term term, boolean observed){
-        this.term = term;
-        this.observed = observed;
-    }
+/**
+ * Encapsulates one (text-mined) HPO term and a boolean to indicate
+ * whether the term was observed or excluded.
+ * @author  Peter N Robinson
+ */
+public record FenominalTerm(Term term, boolean observed)
+        implements Comparable<FenominalTerm> {
 
     public Term getTerm() {
         return term;
@@ -23,18 +19,7 @@ public class FenominalTerm implements Comparable<FenominalTerm> {
         return observed;
     }
 
-    @Override
-    public int hashCode() {
-        return Objects.hash(this.term, this.observed);
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        if (! (obj instanceof FenominalTerm that)) return false;
-        return this.term.equals(that.term) && this.observed == that.observed;
-    }
-
-    public static FenominalTerm fromMainPhenotypeTerm(Main.PhenotypeTerm mpt) {
+    public static FenominalTerm fromMainPhenotypeTerm(PhenotypeTerm mpt) {
         return new FenominalTerm(mpt.getTerm(), mpt.isPresent());
     }
 
