@@ -1,15 +1,22 @@
 package org.monarchinitiative.fenominal.gui.model;
 
 
+import org.monarchinitiative.phenol.ontology.data.Term;
+import org.monarchinitiative.phenol.ontology.data.TermId;
+
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 /**
  * Model for building a Phenopacket with multiple time points for a single patient
  * @author Peter N Robinson
  */
 public class PhenopacketModel implements TextMiningResultsModel {
+
+    private final String id;
 
     private final LocalDate birthdate;
 
@@ -20,8 +27,9 @@ public class PhenopacketModel implements TextMiningResultsModel {
 
 
 
-    public PhenopacketModel(LocalDate bdate) {
+    public PhenopacketModel(LocalDate bdate, String id) {
         birthdate = bdate;
+        this.id = id;
         encounters = new ArrayList<>();
         encounterDates = new ArrayList<>();
     }
@@ -46,12 +54,20 @@ public class PhenopacketModel implements TextMiningResultsModel {
 
     @Override
     public int getTermCount() {
-        return 0;
+        if (encounters == null || encounters.size()==0) return 0;
+        Set<TermId> tids = encounters.stream().map(MedicalEncounter::getTerms)
+                .flatMap(List::stream)
+                .map(FenominalTerm::getTerm)
+                .map(Term::getId)
+                .collect(Collectors.toSet());
+        return tids.size();
     }
 
     public LocalDate getBirthdate() {
         return birthdate;
     }
+
+    public String getPhenopacketId() { return id; }
 
     public List<LocalDate> getEncounterDates() {
         return encounterDates;
